processing functions moved to a module
